Fixed extending the LVRoot volume
If the size of the root volume is set via the @root attribute
in systemdisk and no other allfree volume was specified the
LVRoot volume was extended twice times. first with the specified
size which is ok and later to the full space because no other
volume was specified to take the rest space. The last resize
is unwanted in that case. This Fixes bsc#1060515