Initialize the systemd machine ID on first boot
The systemd and/or dbus machine id is considered to be a unique
information. Thus it is required to initialize it on first boot
of the image. If the image uses the kiwi boot code (initrd)
a hook script handleMachineID is called which allows to perform
the required commands to handle the machine id of the subsystems
which needs this. By default no action takes place because the
setup of the machine id varies depending on the version of the
OS. Thus we leave it up to the user to implement the hook if
needed